+++ Assets/Scripts/Game/Action/Relics/BlackPearlExecute.cs
@@ -51,7 +51,7 @@ namespace Game.Action.Relics
 
         private static Effect CreateEffectFromName(string effectName, PieceLogic piece)
         {
-            var randomDuration = (sbyte)new System.Random().Next(6, 8);
+            var randomDuration = new System.Random().Next(6, 8);
             return effectName switch
             {
                 "effect_shield" => new Effects.Buffs.Shield(piece),
