+++ Assets/Scripts/Game/Effects/Augmentation/ProtectiveLensPassive.cs
@@ -6,7 +6,7 @@ namespace Game.Effects.Augmentation
 {
     public class ProtectiveLensPassive : Effect, IApplyEffect
     {
-        public ProtectiveLensPassive(sbyte duration, sbyte strength, PieceLogic piece) : base(duration, strength, piece, "effect_protected_lens_passive")
+        public ProtectiveLensPassive(int duration, int strength, PieceLogic piece) : base(duration, strength, piece, "effect_protected_lens_passive")
         { }
 
         public void OnCallApplyEffect(ApplyEffect applyEffect)
