+++ Assets/Scripts/Game/Effects/Buffs/Piercing.cs
@@ -7,7 +7,7 @@ namespace Game.Effects.Buffs
     [Il2CppSetOption(Option.NullChecks, false), Il2CppSetOption(Option.ArrayBoundsChecks, false)]
     public class Piercing: Effect, IBeforePieceActionEffect
     {
-        public Piercing(sbyte duration, PieceLogic piece) : base(duration, 1, piece, "effect_piercing")
+        public Piercing(int duration, PieceLogic piece) : base(duration, 1, piece, "effect_piercing")
         {}
 
         public void OnCallBeforePieceAction(Action.Action action)
