+++ Assets/Scripts/Game/Effects/Debuffs/Marked.cs
@@ -4,7 +4,7 @@ namespace Game.Effects.Debuffs
 {
     public class Marked : Effect
     {
-        public Marked(sbyte duration, PieceLogic piece) : base(duration, 1, piece, "effect_marked")
+        public Marked(int duration, PieceLogic piece) : base(duration, 1, piece, "effect_marked")
         {
 
         }
