+++ Assets/Scripts/Game/Piece/PieceConfig.cs
@@ -8,10 +8,10 @@ namespace Game.Piece
     {
         public readonly string Type;
         public readonly bool Color;
-        public readonly ushort Index;
+        public readonly int Index;
         public readonly List<Augmentation.AugmentationName> AugmentationNames;
 
-        public PieceConfig(string type, bool color, ushort index, List<Augmentation.AugmentationName> augmentationNames = null)
+        public PieceConfig(string type, bool color, int index, List<Augmentation.AugmentationName> augmentationNames = null)
         {
             Type = type;
             Color = color;
