+++ Assets/Scripts/Game/Piece/PieceLogic/ArcticBrittleStar.cs
@@ -56,7 +56,7 @@ namespace Game.Piece.PieceLogic
                 }
             };
         }
-        sbyte IPieceWithSkill.TimeToCooldown { get; set; }
+        int IPieceWithSkill.TimeToCooldown { get; set; }
         public SkillsDelegate Skills { get; set; }
     }
