+++ Assets/Scripts/Game/Piece/PieceLogic/BottlenoseDolphin.cs
@@ -36,7 +36,7 @@ namespace Game.Piece.PieceLogic
             };
         }
         
-        sbyte IPieceWithSkill.TimeToCooldown { get; set; }
+        int IPieceWithSkill.TimeToCooldown { get; set; }
         public SkillsDelegate Skills { get; set; }
     }
 }
