+++ Assets/Scripts/Game/Piece/PieceLogic/Commons/IPieceWithSkill.cs
@@ -4,7 +4,7 @@ namespace Game.Piece.PieceLogic.Commons
 {
     public interface IPieceWithSkill
     {
-        protected internal sbyte TimeToCooldown { get; set; }
+        protected internal int TimeToCooldown { get; set; }
         SkillsDelegate Skills { get; }
     }
 }
